Added a base
This base has a WIP Path handler, as well as a proper base to do things such as access Evolve.ipBans from other files. Evolve.js is the main client that handles initiating the paths and initiatizing the base. When the base initiates, it connects to mongodb via a config url or localhost/evolve-x, and then attempts to listen to a port... If it cannot it will error and quit the process. It will also fail if you are trying to use any port below 1024 on linux without root.