Fixed schemas, added ability to add owner. Added Utils
Utils are currently used for authentication and account creation. The first account (The owner account) can be initiated at scripts/initFirst.js. Updated the base file to accomodate for this. Fixed exporting the schemas as well, oops. Added bcrypt dependency for the hashing of tokens and passwords. Bumped version. Added initfirst script