Bug fix, added basic account stuff
Added the ability to have a general account overlook via the api, and the ability to delete your account (unless you are the app owner). Added a little improvement to the Clear Notify path. Made the app work when you run a start command. Updated path docs to go with these new paths.