Updated JSDoc, license, author, & version
Bumped the version as it is obviously no longer at v0.4. Updated the JSDoc because I feel like it could help someone who wants to contribute. Updated the author as I am like the only guy that coded this. I put the license statement in most files so you know when you look at a file that it is in fact under copyright.