Expand what can be uploaded to Evolve-X
Anything, video, file, or image can be uploaded now as long as it is within a 50mb limit. We exclude HTML & JavaScript files from this as they cann screw with stuff, and send a lot of requests back at once unintentionally. This is somewhat of a breaking change. The api images references will be fully changed later on.